feature/refactor-easy-logging
- Refactor EasyLogging
