+++ packages/EasyLogging/src/Bridge/Symfony/Resources/config/definition.php
@@ -1,0 +1,20 @@
+<?php
+declare(strict_types=1);
+
+use EonX\EasyLogging\Interfaces\LoggerFactoryInterface;
+use Symfony\Component\Config\Definition\Configurator\DefinitionConfigurator;
+
+return static function (DefinitionConfigurator $definition) {
+    $definition->rootNode()
+        ->children()
+            ->scalarNode('default_channel')->defaultValue(LoggerFactoryInterface::DEFAULT_CHANNEL)->end()
+            ->arrayNode('sensitive_data_sanitizer')
+                ->addDefaultsIfNotSet()
+                ->children()
+                    ->booleanNode('enabled')->defaultFalse()->end()
+                ->end()
+            ->end()
+            ->booleanNode('stream_handler')->defaultTrue()->end()
+            ->integerNode('stream_handler_level')->defaultNull()->end()
+        ->end();
+};
